ci: add `Test-Corpus` submodule to `Test-Corpus-Mirror`
`Test-Corpus-Mirror` is our private clone of `Test-Corpus`, which is private also,
but for which we can't add a deploy key, hence the private clone.
This adds it as a submodule.
However, since `actions/checkout` can't (IIUC) clone with both HTTPS and SSH,
we update the submodules manually after configuring git
to use the SSH deploy key for `git@gitlab-ext.galois.com`.