analysis: change `cc_cmake_op` to use `bear` to generate `compile_commands.json` instead of `cmake`
`cmake` can generate a `compile_commands.json` without building anything,
but it includes everything, which may include duplicates.
If we build one target with `bear`, then we only get the commands used for that.

For now, this is folded into the existing `cc_cmake_op`
instead of adding a separate `bear_op`.